socketio/requirements.txt
flask
flask-socketio
gevent-websocket
kombu